requirements.txt
Django~=4.2.5
psycopg2-binary~=2.9.7
colorama~=0.4.6
djangorestframework~=3.14.0
django-cors-headers~=4.2.0
Pillow~=10.0.1
uuid~=1.30

# Testing
pytest~=7.4.0
pytest-django~=4.5.0
pytest-cov~=4.1.0
factory-boy~=3.3.0
faker~=19.0.0